refactor code: moved api functions to separate script moved supporting scripts to folder
