+++ src/module/scripts/handleFavoriteControls.js
