+++ src/module/scripts/preloadTemplates.js
