+++ src/module/scripts/settings.js
